% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_find_gene_ids}
\alias{aba_find_gene_ids}
\title{Query the Allen Brain Atlas API for experiment IDs based on gene symbol.}
\usage{
aba_find_gene_ids(gene_symbol, plane = "coronal")
}
\arguments{
\item{gene_symbol}{A character vector containing a gene symbol.}

\item{plane}{The plane of the ISH experiment to search for. Options are "coronal" and "saggital". Default is "coronal".}
}
\value{
a character vector of valid experiment IDs
}
\description{
Query the Allen Brain Atlas API for experiment IDs based on gene symbol.
}
\examples{
ids <- aba_find_gene_ids("Pvalb", "coronal"
)
}
