% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_flatten_ontology}
\alias{aba_flatten_ontology}
\title{Convert the nested Allen Brain Atlas ontology to a data.frame.}
\usage{
aba_flatten_ontology(ontology, ontology_df = NULL)
}
\arguments{
\item{ontology_df}{The current ABA ontology data.frame. Used for recursive unpacking of the list ontology.}

\item{@ontology}{The list-based ontology returned by the aba_get_ontology() function.}
}
\value{
a data.frame containing the ontology, along with a parent_id column for retaining hierarchical information.
}
\description{
Convert the nested Allen Brain Atlas ontology to a data.frame.
}
\examples{
library(dplyr)

ont_df <- aba_get_ontology() \%>\%
  aba_flatten_ontology()

}
