% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_get_annotation_array}
\alias{aba_get_annotation_array}
\title{Retrieve the Allen Brain Atlas Annotation ID array for Gridded expression data.}
\usage{
aba_get_annotation_array()
}
\value{
a 3-dimensional array with dims = c(67, 41, 58)
}
\description{
Retrieve the Allen Brain Atlas Annotation ID array for Gridded expression data.
}
\examples{
anno_arr <- aba_get_annotation_array()

}
