% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_get_energy_array}
\alias{aba_get_energy_array}
\title{Query the Allen Brain Atlas API to retrieve an ISH energy array for an experiment ID}
\usage{
aba_get_energy_array(id)
}
\arguments{
\item{id}{ISH experiment id, like that generated by aba_find_gene_ids()}
}
\value{
A 3-dimensional array of energy values with dims = c(67, 41, 58).
}
\description{
Query the Allen Brain Atlas API to retrieve an ISH energy array for an experiment ID
}
\examples{
library(dplyr)

pvalb_arr <- aba_find_gene_ids("Pvalb") \%>\%
  aba_get_energy_array()

}
