% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_get_ontology}
\alias{aba_get_ontology}
\title{Get the Allen Brain Atlas structure graph ontology from the Allen API}
\usage{
aba_get_ontology()
}
\value{
a nested list containing the allen brain atlas ontology.
}
\description{
Get the Allen Brain Atlas structure graph ontology from the Allen API
}
\examples{
ont <- aba_get_ontology()

}
