% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_melt_2d}
\alias{aba_melt_2d}
\title{Convert an Allen Brain Atlas 2-d array to a data.frame}
\usage{
aba_melt_2d(arr, val = "value", plane = "coronal")
}
\arguments{
\item{arr}{a 2-d array to melt, like that generated by aba_slice_array()}

\item{val}{The name of the value column to use.}

\item{plane}{The plane of the 2-d array. Options are "coronal", "saggital", and "horizontal". Default is "coronal".}
}
\value{
a data.frame with dimension columns based on the plane, and a value column named for val.
}
\description{
Convert an Allen Brain Atlas 2-d array to a data.frame
}
