% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_melt_3d}
\alias{aba_melt_3d}
\title{Convert an Allen Brain Atlas 3-d array to a data.frame}
\usage{
aba_melt_3d(arr, val = "value")
}
\arguments{
\item{arr}{a 3-d array to melt}

\item{val}{A name for the value column. Default is "value".}
}
\value{
a data.frame with columns x, y, z, and val (specified as a parameter).
}
\description{
Convert an Allen Brain Atlas 3-d array to a data.frame
}
