% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_slice_anno_boundaries}
\alias{aba_slice_anno_boundaries}
\title{Generate array annotation boundary segments for ggplot2}
\usage{
aba_slice_anno_boundaries(anno, slice_num, plane = "coronal")
}
\arguments{
\item{anno}{A 3-d annotation array like that generated by aba_get_annotation_array()}

\item{slice_num}{A number indicating which slice to use.}

\item{plane}{Which plane to return. Options are "coronal", "saggital", and "horizontal". Default is "coronal"}
}
\value{
a data.frame of segment coordinates with x, xend, y, and yend.
}
\description{
Generate array annotation boundary segments for ggplot2
}
\examples{
library(dplyr)
library(ggplot2)

boundary_df <- aba_get_annotation_array() \%>\%
  aba_symmetrize_matrix("min") \%>\%
  aba_melt_3d("atlas_id") \%>\%
  filter(atlas_id != 0) \%>\%
  aba_slice_anno_boundaries(30, "coronal")

ggplot(boundary_df) +
  geom_segment(aes(x = x, xend = xend,
                   y = y, yend = yend)) +
  scale_y_reverse()

}
