% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_slice_array}
\alias{aba_slice_array}
\title{Slice an Allen Brain Atlas ISH Array to return a matrix from a single plane.}
\usage{
aba_slice_array(arr, slice, plane = "coronal")
}
\arguments{
\item{arr}{The 3-dimensional ISH array to slice}

\item{slice}{A numeric value for the slice to use.}

\item{plane}{Which plane to return. Options are "coronal", "saggital", and "horizontal". Default is "coronal"}
}
\value{
A 2-dimensional array of ISH values
}
\description{
Slice an Allen Brain Atlas ISH Array to return a matrix from a single plane.
}
\examples{
library(dplyr)

pvalb_mat <- aba_find_gene_ids("Pvalb") \%>\%
  aba_get_energy_array() \%>\%
  aba_slice_array(42, "coronal")

}
