% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_symmetrize_array}
\alias{aba_symmetrize_array}
\title{Apply hemisphere symmetry to an Allen Brain Atlas array}
\usage{
aba_symmetrize_array(arr, fun = "none")
}
\arguments{
\item{arr}{The ISH array to symmetrize}

\item{fun}{The function to use for symmetry calculations. Options are "mean", "max", "min", and "mult".}
}
\value{
a 3-d array with the same dimensions as the input
}
\description{
Apply hemisphere symmetry to an Allen Brain Atlas array
}
