% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{add_grid}
\alias{add_grid}
\title{Add gridding, arrows, and plotting theme for a cross-stitch plot}
\usage{
add_grid(p, x_range = c(0, 100), y_range = c(0, 100), pad = 5,
  center_arrows = TRUE, base_size = 10, base_family = "")
}
\arguments{
\item{p}{The plot object to use.}

\item{x_range}{The range of the x values in the plot. Default is c(0,100), but should be entered per-project.}

\item{y_range}{The range of the y values in the plot. Default is c(0,100), but should be entered per-project.}

\item{pad}{Stitch padding to add around the edges of the pattern. Default is 5.}

\item{center_arrows}{Logical, whether or not to include arrows pointing to the center stitch. Default is TRUE.}

\item{base_size}{Base font size for the plot theme. Default is 10. See ?ggplot2::theme() for more information.}

\item{base_family}{Base font family for the plot theme. Default is "", which will use system defaults.}
}
\value{
A ggplot2 object with grid, optional arrows, and theme applied.
}
\description{
Add gridding, arrows, and plotting theme for a cross-stitch plot
}
\examples{
library(ggplot2)

vmat <- floor((volcano - 100)/30)
vdf <- reshape2::melt(vmat)

p <- ggplot() +
  geom_point(data = vdf,
             aes(x = Var1,
                 y = Var2,
                 pch = as.factor(value),
                 color = as.factor(value)))

add_grid(p,
         x_range = c(1,87),
         y_range = c(1,61),
         center_arrows = T)

}
