% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backstitch.R
\name{backstitch_join}
\alias{backstitch_join}
\title{Join backstitch lettering into a single segment data.frame}
\usage{
backstitch_join(x, anchor = c(0, 0), scale = 1, set = "simple")
}
\arguments{
\item{x}{A character vector containing the lettering to use.}

\item{anchor}{The x,y position of the bottom-left corner of the first letter. Default is c(0,0)}

\item{scale}{Scale of the lettering. I recommend you use integer values to make stitching easier! Default is 1.}

\item{set}{Backstitch set to use. Default is "simple". Use names(backstitch_segments) to see all available.}
}
\value{
A segment data.frame with x, xend, y, and yend for use with geom_segment in ggplot2
}
\description{
Join backstitch lettering into a single segment data.frame
}
\examples{
backstitch_df <- backstitch_join("Easy backstitch phrases!",
                                 anchor = c(-10, 10),
                                 scale = 2,
                                 set = "simple")

head(backstitch_df)

}
