% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Generate a cross-stitch pattern chart
Supply an image and turn it into a cross stitch chart.
Currently assumes full stitches.}
\usage{
chart(imgpath, height = 100, num = NULL, select.col = NULL)
}
\arguments{
\item{imgpath}{the path of the image to be processed}

\item{height}{height of chart to be produced by squares (default is 100)}

\item{num}{number of colours to be used (default is NULL)}

\item{select.col}{provide a vector of coloured strings (DMC numbers/code)}
}
\description{
Generate a cross-stitch pattern chart
Supply an image and turn it into a cross stitch chart.
Currently assumes full stitches.
}
