% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{col2ab}
\alias{col2ab}
\title{Convert colors to alpha-beta values}
\usage{
col2ab(hexes)
}
\arguments{
\item{hexes}{A character vector with a set of hex color values or R colors}
}
\value{
a data.frame with columns for the original color, alpha, and beta values.
}
\description{
See https://en.wikipedia.org/wiki/HSL_and_HSV#Hue_and_chroma for more information.
}
