% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{colorspace_plot}
\alias{colorspace_plot}
\title{Generate a plot in alpha-beta colorspace for a palette}
\usage{
colorspace_plot(palette, show_pures = TRUE)
}
\arguments{
\item{palette}{a character vector containing colors as either hex values (starting with #) or R colors}

\item{show_pures}{a logical value indicating whether or not to plot points for pure colors}
}
\value{
a ggplot2 plot with palette colors in alpha-beta space.
}
\description{
The resulting plot will be a 2-D projection onto the HSV/HSL chromaticity plane.
See https://en.wikipedia.org/wiki/HSL_and_HSV#Hue_and_chroma for more information.
}
