% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cut_array}
\alias{cut_array}
\title{cut() an array and retain dimensions}
\usage{
cut_array(arr, ...)
}
\arguments{
\item{arr}{The array to be processed by cut()}

\item{...}{Arguments passed to the cut() function. See ?cut for details.}
}
\value{
A factor array divided using cut().
}
\description{
cut() an array and retain dimensions
}
\examples{
cut_volcano <- cut_array(volcano, breaks = 5)
}
