% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_launchers.R
\name{launch_aba_shiny}
\alias{launch_aba_shiny}
\title{Launch an instance of the Allen Brain Atlas (ABA) shiny app}
\usage{
launch_aba_shiny()
}
\description{
Launch an instance of the Allen Brain Atlas (ABA) shiny app
}
