% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{nearest_r_color}
\alias{nearest_r_color}
\title{Compute the closest R color for a set of hex colors}
\usage{
nearest_r_color(hexes)
}
\arguments{
\item{hexes}{A character vector of the hex color codes}
}
\value{
A character vector of color names.
}
\description{
This uses colors(distinct = TRUE).
}
