% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backstitch.R
\name{stitch_lettering}
\alias{stitch_lettering}
\title{Add backstitch lettering into a ggplot2 object.}
\usage{
stitch_lettering(p, x, anchor = c(0, 0), scale = 1, set = "simple")
}
\arguments{
\item{p}{The ggplot2 plot to use.}

\item{x}{A character vector with the phrase to add to the plot.}

\item{anchor}{The bottom-left position of the phrase in the plot.}

\item{scale}{Scale of the lettering. I recommend you use integer values to make stitching easier! Default is 1.}

\item{set}{Backstitch set to use. Default is "simple". Use names(backstitch_segments) to see all available.}
}
\value{
A ggplot2 plot object with an added geom_segment() layer.
}
\description{
Add backstitch lettering into a ggplot2 object.
}
\examples{
library(ggplot2)
library(dplyr)

p <- ggplot() +
  geom_point(data = iris,
             aes(x = Sepal.Length,
                 y = Petal.Length,
                 color = Species))

p \%>\%
  stitch_lettering("Nice 150 irises!",
                   anchor = c(4,6),
                   scale = 0.1,
                   set = "simple")

}
