% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{swatch_filter_value}
\alias{swatch_filter_value}
\title{Filter an RGB matrix to remove values with low V in HSV space}
\usage{
swatch_filter_value(swatch_mat, prop = 0.3)
}
\arguments{
\item{swatch_mat}{A 3-column swatch matrix like that generated by extract_swatch()}

\item{prop}{The minimum V value to retain}
}
\value{
A 3-column swatch matrix
}
\description{
This removes darker/dimmer colors
}
