% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{swatch_max_hex}
\alias{swatch_max_hex}
\title{Compute the maximum hex value from an RGB matrix}
\usage{
swatch_max_hex(swatch_mat)
}
\arguments{
\item{swatch_mat}{A 3-column swatch matrix like that generated by extract_swatch()}
}
\value{
A single hex color value
}
\description{
The maximum is based on the maximum of the sums of the separate RGB values.
}
