% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{swatch_median_hex}
\alias{swatch_median_hex}
\title{Compute the median hex value from an RGB matrix}
\usage{
swatch_median_hex(swatch_mat)
}
\arguments{
\item{swatch_mat}{A 3-column swatch matrix like that generated by extract_swatch()}
}
\value{
A single hex color value
}
\description{
The median is based on the median of the sums of the separate RGB values.
}
