% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{swatch_min_hex}
\alias{swatch_min_hex}
\title{Compute the minimum hex value from an RGB matrix}
\usage{
swatch_min_hex(swatch_mat)
}
\arguments{
\item{swatch_mat}{A 3-column swatch matrix like that generated by extract_swatch()}
}
\value{
A single hex color value
}
\description{
The minimum is based on the minimum of the sums of the separate RGB values.
}
