% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatches.R
\name{swatch_most_frequent_hex}
\alias{swatch_most_frequent_hex}
\title{Compute the most frequent hex value from an RGB matrix}
\usage{
swatch_most_frequent_hex(swatch_mat)
}
\arguments{
\item{swatch_mat}{A 3-column swatch matrix like that generated by extract_swatch()}
}
\value{
A single hex color value
}
\description{
Compute the most frequent hex value from an RGB matrix
}
